requirements..txt
langchain
langgraph
chromadb
ollama
pdfplumber
pymupdf
tiktoken
streamlit
tqdm
pytesseract
pillow
pdf2image